refactor: Extract actions to separate files
