+++ apps/client/src/features/page/tree/components/node/components/page-arrow.tsx
@@ -1,0 +1,42 @@
+import { NodeApi } from "react-arborist";
+import { SpaceTreeNode } from "@/features/page/tree/types.ts";
+import { ActionIcon } from "@mantine/core";
+import {
+  IconChevronDown,
+  IconChevronRight,
+  IconPointFilled,
+} from "@tabler/icons-react";
+import React from "react";
+
+interface PageArrowProps {
+  node: NodeApi<SpaceTreeNode>;
+  onExpandTree?: () => void;
+}
+
+export function PageArrow({ node, onExpandTree }: PageArrowProps) {
+  return (
+    <ActionIcon
+      size={20}
+      variant="subtle"
+      c="gray"
+      onClick={(e) => {
+        e.preventDefault();
+        e.stopPropagation();
+        node.toggle();
+        onExpandTree();
+      }}
+    >
+      {node.isInternal ? (
+        node.children && (node.children.length > 0 || node.data.hasChildren) ? (
+          node.isOpen ? (
+            <IconChevronDown stroke={2} size={18} />
+          ) : (
+            <IconChevronRight stroke={2} size={18} />
+          )
+        ) : (
+          <IconPointFilled size={8} />
+        )
+      ) : null}
+    </ActionIcon>
+  );
+}
